augur: split app/ into frontend/ + api/, drop augur_ prefixes
Pure rename refactor. No behavior changes. Splits augur/app/ into:

- augur/frontend/ for browser code: app.jsx (was augur-app.jsx),
  client.js (was augur_client.js), main.jsx, index.html,
  styles.src.css, lib/, __screenshots__/, visual_golden_test.py,
  and all bundle Bazel machinery (esbuild, tailwindcss, bundle_base).
- augur/api/ for server code: backend.py (was augur_backend.py),
  catalog.py, config.py, http_server.py, server.py, export_schema.py,
  browser_state.py (Pydantic SSOT feeding the OpenAPI/Zod codegen),
  browser_shell_test.py (integration test against the server binary),
  and testdata/.

Cross-repo dependency: gaffer-private has parallel updates on
claude/plan-g-augur-app-split that consume the new package paths.
