+++ augur/api/testdata/properties.json
