+++ augur/frontend/__screenshots__/distribution_default.png
