+++ augur/frontend/__screenshots__/distribution_fan.png
