+++ augur/frontend/__screenshots__/trajectory_scenario_2_rollout_3.png
