+++ augur/frontend/lib/backend_client.js
