+++ augur/frontend/lib/casing.js
