+++ augur/frontend/lib/columnar.js
