+++ augur/frontend/lib/scenario_set_state.js
